seguidad de roles y taxonomia

+++ resources/views/clase/index.blade.php
@@ -16,9 +16,11 @@
     <br>
      <div class="row">
        <div class="col-md-8">
+        @can('Crear Clase')
         <button id='btnAgregar' onclick="mostrarFormulario()" class="btn btn-success btn-lg">
             Nuevo Clase
         </button>
+        @endcan
         <br>
         <br>
               {{ Form::open(['route'=>'clase.store', 'method'=>'POST', 'class'=>'agregar']) }}
@@ -59,11 +61,15 @@
         <td>
           <a class="btn btn-info btn-sm" data-toggle="tooltip" data-placement="top" title="Detalles" href="{{route('clase.show',$value->id)}}">
               <i class="glyphicon glyphicon-list-alt">Detalles</i></a>
+          @can('Editar Clase')
           <a class="btn btn-primary btn-sm" data-toggle="tooltip" data-placement="top" title="Editar" href="{{route('clase.edit',$value->id)}}">
               <i class="glyphicon glyphicon-pencil">Editar</i></a>
+          @endcan
+          @can('Eliminar Clase')
             {!! Form::open(['method' => 'DELETE','route' => ['clase.destroy', $value->id],'style'=>'display:inline', 'class'=>'formulario-eliminar']) !!}
               <button type="submit" data-toggle="tooltip" data-placement="top" title="Eliminar" style="display: inline;" class="btn btn-danger btn-sm"><i class="glyphicon glyphicon-trash" >Eliminar</i></button>
             {!! Form::close() !!}
+          @endcan
         </td>
       </tr>
     @endforeach
