+++ resources/views/filum/index.blade.php
@@ -14,9 +14,11 @@
   <br>
      <div class="row">
        <div class="col-md-8">
+        @can('Crear Filum')
         <button id='btnAgregar' onclick="mostrarFormulario()" class="btn btn-success btn-lg">
             Nuevo Filum
         </button>
+        @endcan
         <br>
         <br>
               {{ Form::open(['route'=>'filum.store', 'method'=>'POST', 'class'=>'agregar']) }}
@@ -56,11 +58,15 @@
         <td>
           <a class="btn btn-info btn-sm" data-toggle="tooltip" data-placement="top" title="Detalles" href="{{route('filum.show',$value->id)}}">
               <i class="glyphicon glyphicon-list-alt">Detalles</i></a>
+          @can('Editar Filum')
           <a class="btn btn-primary btn-sm" data-toggle="tooltip" data-placement="top" title="Editar" href="{{route('filum.edit',$value->id)}}">
               <i class="glyphicon glyphicon-pencil">Editar</i></a>
+          @endcan
+          @can('Eliminar Filum')
             {!! Form::open(['method' => 'DELETE','route' => ['filum.destroy', $value->id],'style'=>'display:inline', 'class'=>'formulario-eliminar']) !!}
               <button type="submit" data-toggle="tooltip" data-placement="top" title="Eliminar" style="display: inline;" class="btn btn-danger btn-sm"><i class="glyphicon glyphicon-trash" >Eliminar</i></button>
             {!! Form::close() !!}
+          @endcan
         </td>
       </tr>
     @endforeach
