+++ resources/views/orden/index.blade.php
@@ -15,9 +15,11 @@
   <br>
     <div class="row">
        <div class="col-md-8">
+        @can('Crear Orden')
         <button id='btnAgregar' onclick="mostrarFormulario()" class="btn btn-success btn-lg">
             Nuevo Orden
         </button>
+        @endcan
         <br>
         <br>
               {{ Form::open(['route'=>'orden.store', 'method'=>'POST', 'class'=>'agregar']) }}
@@ -60,11 +62,15 @@
         <td>
           <a class="btn btn-info btn-sm" data-toggle="tooltip" data-placement="top" title="Detalles" href="{{route('orden.show',$value->id)}}">
               <i class="glyphicon glyphicon-list-alt">Detalles</i></a>
+          @can('Editar Orden')
           <a class="btn btn-primary btn-sm" data-toggle="tooltip" data-placement="top" title="Editar" href="{{route('orden.edit',$value->id)}}">
               <i class="glyphicon glyphicon-pencil">Editar</i></a>
+          @endcan
+          @can('Eliminar Orden')
             {!! Form::open(['method' => 'DELETE','route' => ['orden.destroy', $value->id],'style'=>'display:inline', 'class'=>'formulario-eliminar']) !!}
               <button type="submit" data-toggle="tooltip" data-placement="top" title="Eliminar" style="display: inline;" class="btn btn-danger btn-sm"><i class="glyphicon glyphicon-trash" >Eliminar</i></button>
             {!! Form::close() !!}
+          @endcan
         </td>
       </tr>
     @endforeach
